Requested by Aditya: moved draggable and its tests to draggable directory
