+++ src/clr-angular/utils/drag-and-drop/all.spec.ts
@@ -5,10 +5,10 @@
  */
 import ClrDragHandleSpecs from "./drag-handle.spec";
 import ClrDraggableGhostSpecs from "./draggable-ghost.spec";
-import ClrDraggableWithCustomGhostSpecs from "./draggable-with-custom-ghost.spec";
-import ClrDraggableWithDragHandleSpecs from "./draggable-with-drag-handle.spec";
-import ClrDraggableWithGhostAndHandleSpecs from "./draggable-with-ghost-and-handle.spec";
-import ClrDraggableSpecs from "./draggable.spec";
+import ClrDraggableWithCustomGhostSpecs from "./draggable/custom-ghost-integration.spec";
+import ClrDraggableWithDragHandleSpecs from "./draggable/drag-handle-integration.spec";
+import ClrDraggableWithGhostAndHandleSpecs from "./draggable/ghost-and-handle-integration.spec";
+import ClrDraggableSpecs from "./draggable/draggable.spec";
 import ClrIfDraggedSpecs from "./if-dragged.spec";
 import ClrDragAndDropEventBusSpecs from "./providers/drag-and-drop-event-bus.spec";
 import ClrDragEventListenerSpecs from "./providers/drag-event-listener.spec";
@@ -16,7 +16,7 @@ import ClrDragHandleRegistrarSpecs from "./providers/drag-handle-registrar.spec"
 import ClrDraggableSnapshotSpecs from "./providers/draggable-snapshot.spec";
 import ClrGlobalDragModeSpecs from "./providers/global-drag-mode.spec";
 
-fdescribe("Drag And Drop", function() {
+describe("Drag And Drop", function() {
     describe("Providers", function() {
         ClrDragAndDropEventBusSpecs();
         ClrDragEventListenerSpecs();
