+++ src/clr-angular/utils/drag-and-drop/draggable/custom-ghost-integration.spec.ts
@@ -8,15 +8,15 @@ import {TestBed} from "@angular/core/testing";
 import {By} from "@angular/platform-browser";
 import {NoopAnimationsModule} from "@angular/platform-browser/animations";
 
-import {ClrIconModule} from "../../icon/icon.module";
-import {DomAdapter} from "../dom-adapter/dom-adapter";
-import {ClrDragAndDropModule} from "./drag-and-drop.module";
+import {ClrIconModule} from "../../../icon/icon.module";
+import {DomAdapter} from "../../dom-adapter/dom-adapter";
+import {ClrDragAndDropModule} from "../drag-and-drop.module";
 import {ClrDraggable} from "./draggable";
-import {ClrDragEventListener} from "./providers/drag-event-listener";
-import {MOCK_DRAG_EVENT_LISTENER_PROVIDER} from "./providers/drag-event-listener.mock";
-import {ClrDragHandleRegistrar} from "./providers/drag-handle-registrar";
-import {ClrDraggableSnapshot} from "./providers/draggable-snapshot";
-import {ClrGlobalDragMode} from "./providers/global-drag-mode";
+import {ClrDragEventListener} from "../providers/drag-event-listener";
+import {MOCK_DRAG_EVENT_LISTENER_PROVIDER} from "../providers/drag-event-listener.mock";
+import {ClrDragHandleRegistrar} from "../providers/drag-handle-registrar";
+import {ClrDraggableSnapshot} from "../providers/draggable-snapshot";
+import {ClrGlobalDragMode} from "../providers/global-drag-mode";
 
 export default function(): void {
     describe("With Custom Draggable Ghost", function() {
