+++ src/clr-angular/utils/drag-and-drop/draggable/drag-handle-integration.spec.ts
@@ -7,11 +7,11 @@ import {Component} from "@angular/core";
 import {TestBed} from "@angular/core/testing";
 import {By} from "@angular/platform-browser";
 
-import {ClrDragAndDropModule} from "./drag-and-drop.module";
-import {ClrDragHandle} from "./drag-handle";
+import {ClrDragAndDropModule} from "../drag-and-drop.module";
+import {ClrDragHandle} from "../drag-handle";
 import {ClrDraggable} from "./draggable";
-import {ClrDragEventListener} from "./providers/drag-event-listener";
-import {ClrDragHandleRegistrar} from "./providers/drag-handle-registrar";
+import {ClrDragEventListener} from "../providers/drag-event-listener";
+import {ClrDragHandleRegistrar} from "../providers/drag-handle-registrar";
 
 export default function(): void {
     describe("With ClrDragHandle", function() {
